refactor: clean up getEventFromTransaction location
